requirements.txt
dj-database-url==2.3.0
Django==5.1.7
django-heroku==0.3.1
djangorestframework==3.15.2
gunicorn==23.0.0
packaging==24.2
pillow==11.1.0
psycopg2==2.9.10
pysqlite3==0.5.4
python-dotenv==1.1.0
typing_extensions==4.13.0
whitenoise==6.9.0
